requirements.txt
tensorflow>=2.9.0
keras>=2.9.0
numpy>=1.21.0
matplotlib>=3.5.0
tqdm>=4.64.0
pillow>=9.0.0
